Move ownership of ScopedStyleResolver to TreeScope
This removes all the reparenting and tree maintenance logic in favor of
traversing up the tree whenever asking for the parent. If this proves to
be too slow we should just keep a stack of resolvers inside style resolve.

Unfortunately we still need to keep the map for now to allow clean up of
the ScopedStyleResolvers whenever the StyleResolver is discarded to avoid
having free'd CSSStyleSheets in the old resolvers. Eventually we should be
able to fix that by adjusting the lifetime of the sheets.

BUG=392018
TEST=no layout test failures.

Review URL: https://codereview.chromium.org/404363008

git-svn-id: svn://svn.chromium.org/blink/trunk@178795 bbb929c8-8fbe-4397-9dbb-9b2b20218538