Fix RenderLayer::positionFromPaintInvalidationContainer to map the rect to squashing backing correctly.
Previously it was using RenderObject::positionFromPaintInvalidationContainer to map from the paint
invalidation container to the transformed ancestor. This is incorrect, instead it should be using
RenderObject::localToContainerPoint, similar to how RenderLayer::mapRectToPaintBackingCoordinates works.
In practice, it almsot always, or perhaps even always, gets the same result, but the prior state violated
the intent of RenderObject::positionFromPaintInvalidationContainer.

Review URL: https://codereview.chromium.org/402333002

git-svn-id: svn://svn.chromium.org/blink/trunk@178868 bbb929c8-8fbe-4397-9dbb-9b2b20218538