[ServiceWorker] Make fetch() method better conformance with the spec.
In this patch, I don't implement the redirection procedure because it is a bit complicated and it could cause security issues.
When fetch() receives a redirect response, we treat it as an error in DocumentThreadableLoader::redirectReceived().

ResourceResponse's wasFetchedViaServiceWorker flag is used in ResourceFetcher::didReceiveResponse() and DocumentThreadableLoader::handleResponse() for CORS and CSP checking in the renderer side.
But this flag will be set correctly in https://codereview.chromium.org/375513002.
After this patch is landed, I will add LayoutTests for CORS and CSP checking as I described in https://codereview.chromium.org/375513002#msg18. 

BUG=373120

Review URL: https://codereview.chromium.org/399543002

git-svn-id: svn://svn.chromium.org/blink/trunk@178904 bbb929c8-8fbe-4397-9dbb-9b2b20218538