Overlay scrollbars must respect ancestor clip
The current reparenting logic for overlay scrollbars for compositor-
driven scrollable areas does not account for ancestor clips. The result
is that the scrollbars can escape clips that need to affect them.

Consider the following stacking tree. (The '-' characters in the
following trees should be ignored. I've only added them so that
my indenting doesn't get stripped).

+ stacking-context
- + overflow-hidden-clipper
- + overflow-scroller
- + fixed-positioned-descendant-of-the-scroller
- + relative-positioned-descendant-of-the-scroller

Where, in tree order, these are arranged like this
+ stacking-context
- + overflow-hidden-clipper
--- + overflow-scroller
----- + fixed-positioned-descendant-of-the-scroller
----- + relative-positioned-descendant-of-the-scroller

Since overflow-scroller is clipped by overflow-hidden-clipper, a layer
that is not an ancestor in the stacking tree, it gets an ancestor
clipping layer. If the overflow controls are parented under
overflow-scroller's m_graphicsLayer, everything works fine: the
controls inherit the clip from the ancestor clipping layer.

Unfortunately, if we are using overlay scrollbars, we can't hang
the scrollbars under m_graphicsLayer because they'd stack behind
fixed-positioned-descendant-of-the-scroller and
relative-positioned-descendant-of-the-scroller. In order to get things
stacking correctly, we have to reparent the overflow controls like
this:

+ stacking-context
- + overflow-hidden-clipper
- + overflow-scroller
- + fixed-positioned-descendant-of-the-scroller
- + relative-positioned-descendant-of-the-scroller
- + <overflow controls for overflow-scroller>

If we reparent the controls like this in the graphics layer tree,
the overflow controls will no longer be a descendant of the
ancestor clipping layer that applies the clip for m_graphicsLayer,
unfortunately. In order to get the overflow controls clipped
correctly, we need to ensure that they have a clipping layer
above them. This CL adds m_overflowControlsClippingLayer for this
purpose.

BUG=393926

Review URL: https://codereview.chromium.org/397713004

git-svn-id: svn://svn.chromium.org/blink/trunk@178980 bbb929c8-8fbe-4397-9dbb-9b2b20218538