Remove ContainerNode overload for NodeTraversal::nextSkippingChildren()
Remove ContainerNode overload for NodeTraversal::nextSkippingChildren() as it
adds no benefit compared to the one taking a Node in argument.

The nextSkippingChildren() implementation does not involves visiting the
children (and thus calling firstChild() / lastChild()). Therefore, there is no
benefit to knowing that the argument is a ContainerNode (unlike for
NodeTraversal::next()).

This CL also does the same for ElementTraversal::nextSkippingChildren().

R=esprehn@chromium.org, adamk@chromium.org

Review URL: https://codereview.chromium.org/420603012

git-svn-id: svn://svn.chromium.org/blink/trunk@179055 bbb929c8-8fbe-4397-9dbb-9b2b20218538