Get rid of unnecessary isChildNodeList() virtual call in clearChildNodeListCache()
This CL inlines clearChildNodeListCache() in the caller as there is only one to
avoid an unnecessary isChildNodeList() virtual call. The calling method is
invalidateNodeListCachesInAncestors(), which this CL moves from Node to
ContainerNode to make sure it is only called for ContainerNodes (Only
ContainerNodes' node lists have caches that need invalidation).

R=adamk@chromium.org, esprehn@chromium.org

Review URL: https://codereview.chromium.org/413763004

git-svn-id: svn://svn.chromium.org/blink/trunk@179056 bbb929c8-8fbe-4397-9dbb-9b2b20218538