De-flake http/tests/navigation/same-and-different-back.html
The goal of this test is to have two separate iframes navigating
during a back navigation, one same-document navigation and one
different-document navigation. This patch stabilizes the ordering
of actions to guarantee that the creation of a new history entry
happens at the right time.

BUG=355274

Review URL: https://codereview.chromium.org/441723002

git-svn-id: svn://svn.chromium.org/blink/trunk@179534 bbb929c8-8fbe-4397-9dbb-9b2b20218538