Micro-optimization of division in WebGL functions
Avoid division operations in the WebGL API. This is done for uniform*
functions. There size is known to be positive, either by definition or
since it has been checked, which makes it possible to replace
divide-by-POT with right shift. Also, in vertexAttribPointer, modulo is
replaced by bitwise & since the sizes of all the GL types are known to be
powers of two.

Also remove a redundant check from vertexAttribPointer, running the code
with asserts enabled should be enough to ensure that the check is never
hit.

Performance improvement around 2% was seen in one renderer thread limited
test case which particularly stressed uniform* functions.

BUG=400135
TEST=WebGL conformance tests

Review URL: https://codereview.chromium.org/428123005

git-svn-id: svn://svn.chromium.org/blink/trunk@179631 bbb929c8-8fbe-4397-9dbb-9b2b20218538