Properly handle zero height lines with image shape-outside
When a zero height line is supplied and the image shape extends into the
margin box (only possible when a shape-margin is supplied), then only an
empty interval was being returned. This patch makes it properly return
the interval for the line in question.

BUG=397728

Review URL: https://codereview.chromium.org/434623009

git-svn-id: svn://svn.chromium.org/blink/trunk@179642 bbb929c8-8fbe-4397-9dbb-9b2b20218538