FetchEvent should not forcibly reject an unresolved Promise when destructed
Currently an unresolved Promise on RespondWithObserver is forcibly rejected
when the RespondWithObserver is destructed. This means that an unresolved Promise
registered with fetchEvent.respondWith() is forcibly rejected when the fetchEvent
is destructed. This behavior is not good because it makes GC observable.
(For more details, see comment #2 and #4 in the CL discussion.)

Given the above, this CL stops rejecting an unresolved Promise in
RespondWithObserver's destructor. Also this CL removes a test that were
testing the behavior. The removed test should hang after this change
(because an unresolved Promise is never resolved nor rejected).

BUG=400645

Review URL: https://codereview.chromium.org/442183003

git-svn-id: svn://svn.chromium.org/blink/trunk@179695 bbb929c8-8fbe-4397-9dbb-9b2b20218538