Add --write-full-results-to for compatibility w/ other python harnesses.
Eventually we'll want run-webkit-tests to reuse all of the json result
generating and uploading code that the other harnesses will use. This
at least will ensure we support the same command line flags so that we
can make this be a PythonBasedTest in the chromium.Steps in build recipes.

R=ojan@chromium.org, ivandavid@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/451523003

git-svn-id: svn://svn.chromium.org/blink/trunk@179720 bbb929c8-8fbe-4397-9dbb-9b2b20218538