Implement ScriptPromiseProperty::reset
Sometimes we want to replace a promise property of an object.
Because the internal state is stored in each holder wrapper and the destruction
timing is not deterministic, we should have an explicit reset method.
The method re-initializes the ScriptPromiseProperty object so that
promise() returns a new Promise object and you can resolve or reject
the promise without affecting the previously issued Promise.

BUG=400613

Review URL: https://codereview.chromium.org/445453002

git-svn-id: svn://svn.chromium.org/blink/trunk@179815 bbb929c8-8fbe-4397-9dbb-9b2b20218538